python/pyproject.toml
[project]
name = "kumiho"
version = "0.2.0"
license = { file = "LICENSE" }
dependencies = [
	"grpcio>=1.76.0",
	"grpcio-tools>=1.76.0",
	"protobuf>=6.31.1",
	"requests>=2.31.0",
]

[project.scripts]
kumiho-auth = "kumiho.auth_cli:main"

[project.optional-dependencies]
dev = [
	"pytest>=8.2.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"